use action class
